Refactored Home component into a SPA architecture,encapsulated Skill and Project components within the Modal component
