+++ src/components/Modal/Modal.css
@@ -1,11 +1,15 @@
+.modals {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
 .folders {
   display: grid;
   grid-template-columns: repeat(4, 1fr);
   grid-gap: 15px;
   margin: 0 auto;
-  margin-top: 30px;
   width: 90vw;
-  margin-bottom: 10vh;
 }
 
 .btn-modal {
@@ -20,11 +24,10 @@
 
 .btn-modal h3{
   font-weight: 100;
-  /* font-size: 19px; */
 }
 
 .modal, .overlay {
-  width: 100vw;
+  width: 100%;
   height: 100vh;
   top: 0;
   left: 0;
