+++ src/components/Modal/Modal.js
@@ -20,7 +20,7 @@ function Modal() {
   }
 
   return (
-    <>
+    <div className='modals' id='files'>
       <div className='folders'>
         <div onClick={toggleSkillModal} className='btn-modal'>
           <img rel="preload" src='assets/icons/folder.png' alt='folder' style={{width: '48px', height: '48px'}} />
@@ -66,10 +66,6 @@ function Modal() {
               </div>
           </div>)
         }
-        {/* <a href="assets/resume.pdf" target='_blank' rel="preload noreferrer noopener" className='btn-modal' style={{textDecoration: 'none', color: 'black'}}>
-          <img rel="preload" src="assets/icons/notepad.png" alt="resume" style={{width: '48px'}} />
-          <p>Resume</p>
-        </a> */}
         <a href="https://www.linkedin.com/in/hanbyeol-kwon/" target='_blank' rel="noreferrer noopener" className='btn-modal' style={{textDecoration: 'none', color: 'black'}}>
           <img rel="preload" src="assets/icons/linkedin.png" alt="linkedin" style={{width: '48px', height: '48px'}} />
           <p>LinkedIn</p>
@@ -78,12 +74,8 @@ function Modal() {
           <img rel="preload" src="assets/icons/github.png" alt="github" style={{width: '48px', height: '48px'}} />
           <p>Github</p>
         </a>
-        <a href="/contact" className='btn-modal btn-gh' style={{textDecoration: 'none', color: 'black'}}>
-          <img rel="preload" src="assets/icons/email.png" alt="contact" style={{width: '48px', height: '48px'}} />
-          <p>Contact</p>
-        </a>
       </div>
-    </>
+    </div>
   );
 }
