+++ src/components/Navbar/Navbar.js
@@ -1,5 +1,4 @@
 import { useState } from 'react';
-import { Link } from 'react-router-dom';
 import Dropdown from '../Dropdown/Dropdown';
 import './Navbar.css';
 import Time from '../Time/Time.js';
@@ -25,10 +24,10 @@ function Navbar() {
       <nav className='navbar'>
         <div className='navbar-container'>
             <div className='nav-item' onMouseEnter={onMouseEnter} onMouseLeave={onMouseLeave} onClick={onClick}>
-              <Link className={dropdown ? 'nav-hover' : 'nav-links'}>
+              <div className={dropdown ? 'nav-hover' : 'nav-links'}>
                 <img src='assets/icons/logo-icon.png?as=webp' alt='Logo' className='navbar-logo' style={{width: '19px', height: '19px'}} />
                 Start
-              </Link>
+              </div>
               {dropdown && <Dropdown />}
             </div>
             <div className='nav-item'><Time /></div>
