+++ src/components/pages/Home.css
@@ -1,0 +1,25 @@
+.home {
+  height: 100vh;
+}
+
+.sections {
+  background-image: url('../../../public/assets/bg.jpg');
+  background-size: cover;
+  width: 100%;
+  height: calc(100vh - 40px);
+  position: relative;
+  overflow-x: hidden;
+  scroll-snap-type: y mandatory;
+  scroll-behavior: smooth;
+  scrollbar-width: none;
+}
+
+.sections::-webkit-scrollbar{
+  display: none;
+}
+
+.sections > * {
+  width: 100%;
+  height: calc(100vh - 40px);
+  scroll-snap-align: start;
+}
